Minor fix to event detection mechanism

+++ apps/sensorsunleashed/susensorcommon.h
@@ -45,6 +45,6 @@
 int suconfig(struct susensors_sensor* this, int type, void* data);
 void setResource(struct susensors_sensor* this, resource_t* res);
 void setEventU8(struct susensors_sensor* this, int dir, uint8_t step);
-void setEventU16(struct susensors_sensor* this, int dir, uint8_t step);
-void setEventU32(struct susensors_sensor* this, int dir, uint8_t step);
+void setEventU16(struct susensors_sensor* this, int dir, uint16_t step);
+void setEventU32(struct susensors_sensor* this, int dir, uint32_t step);
 #endif /* SENSORSUNLEASHED_DEV_SUSENSORCOMMON_H_ */
